pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.12"
strict = true
exclude = [
  "^.*_gunicorn_config\\.py$"
]

[tool.poetry]
name = "incrementum"
version = "0.1.0"
description = ""
authors = ["Andrew Sheridan <andrewasheridan@gmail.com>"]
readme = "README.md"
packages = [{include = "incrementum_api", from = "app"}]

[tool.poetry.dependencies]
python = "^3.12,<4.0"
litestar = {version = "^2.4.5", extras = ["standard", "pydantic", "redis"]}
uvicorn = {extras = ["standard"], version = "^0.25.0"}
pydantic-settings = "^2.1.0"
gunicorn = "^21.2.0"
exemplars = {git = "https://github.com/andrewasheridan/exemplars.git", branch = "main"}

[tool.poetry.group.dev.dependencies]
setuptools = "^69.0.3"
ruff = "^0.1.9"
pre-commit = "^3.6.0"
mypy = "^1.8.0"
types-redis = "^4.6.0.11"

[tool.ruff]
line-length = 120
indent-width = 4
fix = true
target-version = "py312"
src = ["app", "test"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
extend-select = ["D", "I"]
extend-ignore = ["D203", "D212"]

[tool.ruff.lint.isort]
known-first-party = ["incrementum_api", "exemplars"]

[tool.ruff.lint.pydocstyle]
convention = "google"
